move unused xquartz.rb to new namespace
* migrate methods app_with_bundle_id and pkgutil_info within xquartz.rb
  as this is the only place they are used
* remove duplicate X11 namespace
