+++ lib/homebrew-fork/Library/Homebrew/os/mac.rb
@@ -1,6 +1,5 @@
 require 'hardware'
 require 'os/mac/version'
-require 'os/mac/xquartz'
 
 module OS
   module Mac
@@ -16,17 +15,6 @@ module OS
 
     def prefer_64_bit?
       Hardware::CPU.is_64_bit? and version > :leopard
-    end
-
-    def app_with_bundle_id(*ids)
-      path = mdfind(*ids).first
-      Pathname.new(path) unless path.nil? or path.empty?
-    end
-
-    def pkgutil_info(id)
-      (@pkginfo ||= {}).fetch(id) do |key|
-        @pkginfo[key] = Utils.popen_read("/usr/sbin/pkgutil", "--pkg-info", key).strip
-      end
     end
   end
 end
